+++ ashley/src/test/kotlin/ktx/ashley/ComponentTagSpec.kt
@@ -7,49 +7,102 @@ import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
 
-class TagComponent(val v: Int = 0) : Component
-class ComponentWithoutDefault(val v: Int) : Component
+class FlagComponent : Component
 
-var Entity.defaultConstructorTag by tagFor<TagComponent>()
-var Entity.singletonTag by tagFor(TagComponent(1))
-var Entity.providerTag by tagFor<TagComponent> { TagComponent(42) }
-val Entity.readOnlyTag by tagFor<ComponentWithoutDefault>()
+var Entity.singletonTagCreatedByReflection by tagFor<FlagComponent>()
+var Entity.providerTagCreatedByReflection by tagFor<FlagComponent>(singleton = false)
+var Entity.singletonTagCreatedWithProvider by tagFor<FlagComponent> { FlagComponent() }
+var Entity.providerTagCreatedWithProvider by tagFor(singleton = false) { FlagComponent() }
+var Entity.singletonTagCreatedWithInstance by tagFor(FlagComponent())
 
 object ComponentTagSpec : Spek({
-
-  describe("utilities for tag component properties") {
+  describe("utilities tag component delegates") {
     val entity = Entity()
+    val mapper = mapperFor<FlagComponent>()
 
-    it("Default ConstructorTag") {
-      assertThat(entity.defaultConstructorTag).isFalse()
-      entity.defaultConstructorTag = true
-      assertThat(entity.defaultConstructorTag).isTrue()
-      assertThat(entity.getComponent(TagComponent::class.java).v).isEqualTo(0)
-      entity.defaultConstructorTag = false
-      assertThat(entity.defaultConstructorTag).isFalse()
+    it("should detect if component is not defined with singleton tag created by reflection") {
+      assertThat(entity.singletonTagCreatedByReflection).isFalse()
     }
-    it("Singleton Tag") {
-      assertThat(entity.singletonTag).isFalse()
-      entity.singletonTag = true
-      assertThat(entity.singletonTag).isTrue()
-      assertThat(entity.getComponent(TagComponent::class.java).v).isEqualTo(1)
-      entity.singletonTag = false
-      assertThat(entity.singletonTag).isFalse()
+    it("should detect if component is defined with singleton tag created by reflection") {
+      entity.add(FlagComponent())
+      assertThat(entity.singletonTagCreatedByReflection).isTrue()
     }
-    it("Provider Tag") {
-      assertThat(entity.providerTag).isFalse()
-      entity.providerTag = true
-      assertThat(entity.providerTag).isTrue()
-      assertThat(entity.getComponent(TagComponent::class.java).v).isEqualTo(42)
-      entity.providerTag = false
-      assertThat(entity.providerTag).isFalse()
+    it("should add component with singleton tag created by reflection") {
+      entity.singletonTagCreatedByReflection = true
+      assertThat(entity.has(mapper)).isTrue()
     }
-    it("ReadOnly Tag") {
-      assertThat(entity.readOnlyTag).isFalse()
-      entity.add(ComponentWithoutDefault(1))
-      assertThat(entity.readOnlyTag).isTrue()
-      entity.remove(ComponentWithoutDefault::class.java)
-      assertThat(entity.readOnlyTag).isFalse()
+    it("should remove component with singleton tag created by reflection") {
+      entity.add(FlagComponent())
+      entity.singletonTagCreatedByReflection = false
+      assertThat(entity.has(mapper)).isFalse()
+    }
+
+    it("should detect if component is not defined with provider tag created by reflection") {
+      assertThat(entity.providerTagCreatedByReflection).isFalse()
+    }
+    it("should detect if component is defined with provider tag created by reflection") {
+      entity.add(FlagComponent())
+      assertThat(entity.providerTagCreatedByReflection).isTrue()
+    }
+    it("should add component with provider tag created by reflection") {
+      entity.providerTagCreatedByReflection = true
+      assertThat(entity.has(mapper)).isTrue()
+    }
+    it("should remove component with provider tag created by reflection") {
+      entity.add(FlagComponent())
+      entity.providerTagCreatedByReflection = false
+      assertThat(entity.has(mapper)).isFalse()
+    }
+
+    it("should detect if component is not defined with singleton tag created with provider") {
+      assertThat(entity.singletonTagCreatedWithProvider).isFalse()
+    }
+    it("should detect if component is defined with singleton tag created with provider") {
+      entity.add(FlagComponent())
+      assertThat(entity.singletonTagCreatedWithProvider).isTrue()
+    }
+    it("should add component with singleton tag created with provider") {
+      entity.singletonTagCreatedWithProvider = true
+      assertThat(entity.has(mapper)).isTrue()
+    }
+    it("should remove component with singleton tag created with provider") {
+      entity.add(FlagComponent())
+      entity.singletonTagCreatedWithProvider = false
+      assertThat(entity.has(mapper)).isFalse()
+    }
+
+    it("should detect if component is not defined with provider tag created with provider") {
+      assertThat(entity.providerTagCreatedWithProvider).isFalse()
+    }
+    it("should detect if component is defined with provider tag created with provider") {
+      entity.add(FlagComponent())
+      assertThat(entity.providerTagCreatedWithProvider).isTrue()
+    }
+    it("should add component with provider tag created with provider") {
+      entity.providerTagCreatedWithProvider = true
+      assertThat(entity.has(mapper)).isTrue()
+    }
+    it("should remove component with provider tag created with provider") {
+      entity.add(FlagComponent())
+      entity.providerTagCreatedWithProvider = false
+      assertThat(entity.has(mapper)).isFalse()
+    }
+
+    it("should detect if component is not defined with singleton tag created with instance") {
+      assertThat(entity.singletonTagCreatedWithInstance).isFalse()
+    }
+    it("should detect if component is defined with singleton tag created with instance") {
+      entity.add(FlagComponent())
+      assertThat(entity.singletonTagCreatedWithInstance).isTrue()
+    }
+    it("should add component with singleton tag created with instance") {
+      entity.singletonTagCreatedWithInstance = true
+      assertThat(entity.has(mapper)).isTrue()
+    }
+    it("should remove component with singleton tag created with instance") {
+      entity.add(FlagComponent())
+      entity.singletonTagCreatedWithInstance = false
+      assertThat(entity.has(mapper)).isFalse()
     }
   }
 })